Move point to new pkg & add readme file
