--- pkg/primitives/point.go
@@ -1,10 +1,0 @@
-package primitives
-
-type Point struct {
-	Coordinate Vector
-	Color      Vector
-}
-
-func (p *Point) SetColor(color Vector) {
-	p.Color = color
-}
