+++ pkg/primitives/point/README.md
@@ -1,0 +1,11 @@
+# Point
+
+It represents a point, so that this structure contains 2 vectors. one for the position (Coordinate), and one for the color (Color).
+
+## Functions
+
+It has only one setter function.
+
+## SetColor
+
+It updates the `Color` of the Point.
